requirements.txt
stravalib
supabase
python-dotenv
fastapi
uvicorn[standard]
click
